move vagrant files into test dir
