+++ tests/minikubeOnVagrant/README.txt
