+++ tests/minikubeOnVagrant/Vagrantfile
